+++ sdk/hybridcompute/azure-resourcemanager-hybridcompute/src/main/java/com/azure/resourcemanager/hybridcompute/models/PatchServiceUsed.java
@@ -5,7 +5,6 @@
 package com.azure.resourcemanager.hybridcompute.models;
 
 import com.azure.core.util.ExpandableStringEnum;
-import com.fasterxml.jackson.annotation.JsonCreator;
 import java.util.Collection;
 
 /**
@@ -57,7 +56,6 @@ public final class PatchServiceUsed extends ExpandableStringEnum<PatchServiceUse
      * @param name a name to look for.
      * @return the corresponding PatchServiceUsed.
      */
-    @JsonCreator
     public static PatchServiceUsed fromString(String name) {
         return fromString(name, PatchServiceUsed.class);
     }