+++ sdk/resourcemanager/azure-resourcemanager-compute/src/main/java/com/azure/resourcemanager/compute/models/StorageProfile.java
@@ -30,20 +30,19 @@ public final class StorageProfile {
     private OSDisk osDisk;
 
     /*
-     * Specifies the parameters that are used to add a data disk to a virtual machine. For more information about
-     * disks, see [About disks and VHDs for Azure virtual
+     * Specifies the parameters that are used to add a data disk to a virtual machine. For more information about disks,
+     * see [About disks and VHDs for Azure virtual
      * machines](https://docs.microsoft.com/azure/virtual-machines/managed-disks-overview).
      */
     @JsonProperty(value = "dataDisks")
     private List<DataDisk> dataDisks;
 
     /*
-     * Specifies the disk controller type configured for the VM. **Note:** This property will be set to the default
-     * disk controller type if not specified provided virtual machine is being created with 'hyperVGeneration' set to
-     * V2 based on the capabilities of the operating system disk and VM size from the the specified minimum api
-     * version. You need to deallocate the VM before updating its disk controller type unless you are updating the VM
-     * size in the VM configuration which implicitly deallocates and reallocates the VM. Minimum api-version:
-     * 2022-08-01.
+     * Specifies the disk controller type configured for the VM. **Note:** This property will be set to the default disk
+     * controller type if not specified provided virtual machine is being created with 'hyperVGeneration' set to V2
+     * based on the capabilities of the operating system disk and VM size from the the specified minimum api version.
+     * You need to deallocate the VM before updating its disk controller type unless you are updating the VM size in the
+     * VM configuration which implicitly deallocates and reallocates the VM. Minimum api-version: 2022-08-01.
      */
     @JsonProperty(value = "diskControllerType")
     private DiskControllerTypes diskControllerType;
@@ -129,10 +128,10 @@ public final class StorageProfile {
     /**
      * Get the diskControllerType property: Specifies the disk controller type configured for the VM. **Note:** This
      * property will be set to the default disk controller type if not specified provided virtual machine is being
-     * created with 'hyperVGeneration' set to V2 based on the capabilities of the operating system disk and VM size
-     * from the the specified minimum api version. You need to deallocate the VM before updating its disk controller
-     * type unless you are updating the VM size in the VM configuration which implicitly deallocates and reallocates
-     * the VM. Minimum api-version: 2022-08-01.
+     * created with 'hyperVGeneration' set to V2 based on the capabilities of the operating system disk and VM size from
+     * the the specified minimum api version. You need to deallocate the VM before updating its disk controller type
+     * unless you are updating the VM size in the VM configuration which implicitly deallocates and reallocates the VM.
+     * Minimum api-version: 2022-08-01.
      * 
      * @return the diskControllerType value.
      */
@@ -143,10 +142,10 @@ public final class StorageProfile {
     /**
      * Set the diskControllerType property: Specifies the disk controller type configured for the VM. **Note:** This
      * property will be set to the default disk controller type if not specified provided virtual machine is being
-     * created with 'hyperVGeneration' set to V2 based on the capabilities of the operating system disk and VM size
-     * from the the specified minimum api version. You need to deallocate the VM before updating its disk controller
-     * type unless you are updating the VM size in the VM configuration which implicitly deallocates and reallocates
-     * the VM. Minimum api-version: 2022-08-01.
+     * created with 'hyperVGeneration' set to V2 based on the capabilities of the operating system disk and VM size from
+     * the the specified minimum api version. You need to deallocate the VM before updating its disk controller type
+     * unless you are updating the VM size in the VM configuration which implicitly deallocates and reallocates the VM.
+     * Minimum api-version: 2022-08-01.
      * 
      * @param diskControllerType the diskControllerType value to set.
      * @return the StorageProfile object itself.