+++ sdk/synapse/azure-analytics-synapse-accesscontrol/src/main/java/com/azure/analytics/synapse/accesscontrol/models/RoleAssignmentDetails.java
@@ -5,45 +5,53 @@
 package com.azure.analytics.synapse.accesscontrol.models;
 
 import com.azure.core.annotation.Fluent;
-import com.fasterxml.jackson.annotation.JsonProperty;
+import com.azure.json.JsonReader;
+import com.azure.json.JsonSerializable;
+import com.azure.json.JsonToken;
+import com.azure.json.JsonWriter;
+import java.io.IOException;
+import java.util.Objects;
 import java.util.UUID;
 
-/** Role Assignment response details. */
+/**
+ * Role Assignment response details.
+ */
 @Fluent
-public final class RoleAssignmentDetails {
+public final class RoleAssignmentDetails implements JsonSerializable<RoleAssignmentDetails> {
     /*
      * Role Assignment ID
      */
-    @JsonProperty(value = "id")
     private String id;
 
     /*
      * Role ID of the Synapse Built-In Role
      */
-    @JsonProperty(value = "roleDefinitionId")
     private UUID roleDefinitionId;
 
     /*
      * Object ID of the AAD principal or security-group
      */
-    @JsonProperty(value = "principalId")
     private UUID principalId;
 
     /*
      * Scope at the role assignment is created
      */
-    @JsonProperty(value = "scope")
     private String scope;
 
     /*
      * Type of the principal Id: User, Group or ServicePrincipal
      */
-    @JsonProperty(value = "principalType")
     private String principalType;
 
     /**
+     * Creates an instance of RoleAssignmentDetails class.
+     */
+    public RoleAssignmentDetails() {
+    }
+
+    /**
      * Get the id property: Role Assignment ID.
-     *
+     * 
      * @return the id value.
      */
     public String getId() {
@@ -52,7 +60,7 @@ public final class RoleAssignmentDetails {
 
     /**
      * Set the id property: Role Assignment ID.
-     *
+     * 
      * @param id the id value to set.
      * @return the RoleAssignmentDetails object itself.
      */
@@ -63,7 +71,7 @@ public final class RoleAssignmentDetails {
 
     /**
      * Get the roleDefinitionId property: Role ID of the Synapse Built-In Role.
-     *
+     * 
      * @return the roleDefinitionId value.
      */
     public UUID getRoleDefinitionId() {
@@ -72,7 +80,7 @@ public final class RoleAssignmentDetails {
 
     /**
      * Set the roleDefinitionId property: Role ID of the Synapse Built-In Role.
-     *
+     * 
      * @param roleDefinitionId the roleDefinitionId value to set.
      * @return the RoleAssignmentDetails object itself.
      */
@@ -83,7 +91,7 @@ public final class RoleAssignmentDetails {
 
     /**
      * Get the principalId property: Object ID of the AAD principal or security-group.
-     *
+     * 
      * @return the principalId value.
      */
     public UUID getPrincipalId() {
@@ -92,7 +100,7 @@ public final class RoleAssignmentDetails {
 
     /**
      * Set the principalId property: Object ID of the AAD principal or security-group.
-     *
+     * 
      * @param principalId the principalId value to set.
      * @return the RoleAssignmentDetails object itself.
      */
@@ -103,7 +111,7 @@ public final class RoleAssignmentDetails {
 
     /**
      * Get the scope property: Scope at the role assignment is created.
-     *
+     * 
      * @return the scope value.
      */
     public String getScope() {
@@ -112,7 +120,7 @@ public final class RoleAssignmentDetails {
 
     /**
      * Set the scope property: Scope at the role assignment is created.
-     *
+     * 
      * @param scope the scope value to set.
      * @return the RoleAssignmentDetails object itself.
      */
@@ -123,7 +131,7 @@ public final class RoleAssignmentDetails {
 
     /**
      * Get the principalType property: Type of the principal Id: User, Group or ServicePrincipal.
-     *
+     * 
      * @return the principalType value.
      */
     public String getPrincipalType() {
@@ -132,12 +140,62 @@ public final class RoleAssignmentDetails {
 
     /**
      * Set the principalType property: Type of the principal Id: User, Group or ServicePrincipal.
-     *
+     * 
      * @param principalType the principalType value to set.
      * @return the RoleAssignmentDetails object itself.
      */
     public RoleAssignmentDetails setPrincipalType(String principalType) {
         this.principalType = principalType;
         return this;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public JsonWriter toJson(JsonWriter jsonWriter) throws IOException {
+        jsonWriter.writeStartObject();
+        jsonWriter.writeStringField("id", this.id);
+        jsonWriter.writeStringField("roleDefinitionId", Objects.toString(this.roleDefinitionId, null));
+        jsonWriter.writeStringField("principalId", Objects.toString(this.principalId, null));
+        jsonWriter.writeStringField("scope", this.scope);
+        jsonWriter.writeStringField("principalType", this.principalType);
+        return jsonWriter.writeEndObject();
+    }
+
+    /**
+     * Reads an instance of RoleAssignmentDetails from the JsonReader.
+     * 
+     * @param jsonReader The JsonReader being read.
+     * @return An instance of RoleAssignmentDetails if the JsonReader was pointing to an instance of it, or null if it
+     * was pointing to JSON null.
+     * @throws IOException If an error occurs while reading the RoleAssignmentDetails.
+     */
+    public static RoleAssignmentDetails fromJson(JsonReader jsonReader) throws IOException {
+        return jsonReader.readObject(reader -> {
+            RoleAssignmentDetails deserializedRoleAssignmentDetails = new RoleAssignmentDetails();
+            while (reader.nextToken() != JsonToken.END_OBJECT) {
+                String fieldName = reader.getFieldName();
+                reader.nextToken();
+
+                if ("id".equals(fieldName)) {
+                    deserializedRoleAssignmentDetails.id = reader.getString();
+                } else if ("roleDefinitionId".equals(fieldName)) {
+                    deserializedRoleAssignmentDetails.roleDefinitionId
+                        = reader.getNullable(nonNullReader -> UUID.fromString(nonNullReader.getString()));
+                } else if ("principalId".equals(fieldName)) {
+                    deserializedRoleAssignmentDetails.principalId
+                        = reader.getNullable(nonNullReader -> UUID.fromString(nonNullReader.getString()));
+                } else if ("scope".equals(fieldName)) {
+                    deserializedRoleAssignmentDetails.scope = reader.getString();
+                } else if ("principalType".equals(fieldName)) {
+                    deserializedRoleAssignmentDetails.principalType = reader.getString();
+                } else {
+                    reader.skipChildren();
+                }
+            }
+
+            return deserializedRoleAssignmentDetails;
+        });
     }
 }