+++ sdk/synapse/azure-analytics-synapse-monitoring/src/main/java/com/azure/analytics/synapse/monitoring/implementation/MonitoringClientImpl.java
@@ -6,68 +6,79 @@ package com.azure.analytics.synapse.monitoring.implementation;
 
 import com.azure.core.http.HttpPipeline;
 import com.azure.core.http.HttpPipelineBuilder;
-import com.azure.core.http.policy.CookiePolicy;
 import com.azure.core.http.policy.RetryPolicy;
 import com.azure.core.http.policy.UserAgentPolicy;
 import com.azure.core.util.serializer.JacksonAdapter;
 import com.azure.core.util.serializer.SerializerAdapter;
 
-/** Initializes a new instance of the MonitoringClient type. */
+/**
+ * Initializes a new instance of the MonitoringClient type.
+ */
 public final class MonitoringClientImpl {
-    /** The workspace development endpoint, for example https://myworkspace.dev.azuresynapse.net. */
+    /**
+     * The workspace development endpoint, for example https://myworkspace.dev.azuresynapse.net.
+     */
     private final String endpoint;
 
     /**
      * Gets The workspace development endpoint, for example https://myworkspace.dev.azuresynapse.net.
-     *
+     * 
      * @return the endpoint value.
      */
     public String getEndpoint() {
         return this.endpoint;
     }
 
-    /** Api Version. */
+    /**
+     * Api Version.
+     */
     private final String apiVersion;
 
     /**
      * Gets Api Version.
-     *
+     * 
      * @return the apiVersion value.
      */
     public String getApiVersion() {
         return this.apiVersion;
     }
 
-    /** The HTTP pipeline to send requests through. */
+    /**
+     * The HTTP pipeline to send requests through.
+     */
     private final HttpPipeline httpPipeline;
 
     /**
      * Gets The HTTP pipeline to send requests through.
-     *
+     * 
      * @return the httpPipeline value.
      */
     public HttpPipeline getHttpPipeline() {
         return this.httpPipeline;
     }
 
-    /** The serializer to serialize an object into a string. */
+    /**
+     * The serializer to serialize an object into a string.
+     */
     private final SerializerAdapter serializerAdapter;
 
     /**
      * Gets The serializer to serialize an object into a string.
-     *
+     * 
      * @return the serializerAdapter value.
      */
     public SerializerAdapter getSerializerAdapter() {
         return this.serializerAdapter;
     }
 
-    /** The MonitoringsImpl object to access its operations. */
+    /**
+     * The MonitoringsImpl object to access its operations.
+     */
     private final MonitoringsImpl monitorings;
 
     /**
      * Gets the MonitoringsImpl object to access its operations.
-     *
+     * 
      * @return the MonitoringsImpl object.
      */
     public MonitoringsImpl getMonitorings() {
@@ -76,23 +87,18 @@ public final class MonitoringClientImpl {
 
     /**
      * Initializes an instance of MonitoringClient client.
-     *
+     * 
      * @param endpoint The workspace development endpoint, for example https://myworkspace.dev.azuresynapse.net.
      * @param apiVersion Api Version.
      */
     public MonitoringClientImpl(String endpoint, String apiVersion) {
-        this(
-                new HttpPipelineBuilder()
-                        .policies(new UserAgentPolicy(), new RetryPolicy(), new CookiePolicy())
-                        .build(),
-                JacksonAdapter.createDefaultSerializerAdapter(),
-                endpoint,
-                apiVersion);
+        this(new HttpPipelineBuilder().policies(new UserAgentPolicy(), new RetryPolicy()).build(),
+            JacksonAdapter.createDefaultSerializerAdapter(), endpoint, apiVersion);
     }
 
     /**
      * Initializes an instance of MonitoringClient client.
-     *
+     * 
      * @param httpPipeline The HTTP pipeline to send requests through.
      * @param endpoint The workspace development endpoint, for example https://myworkspace.dev.azuresynapse.net.
      * @param apiVersion Api Version.
@@ -103,14 +109,14 @@ public final class MonitoringClientImpl {
 
     /**
      * Initializes an instance of MonitoringClient client.
-     *
+     * 
      * @param httpPipeline The HTTP pipeline to send requests through.
      * @param serializerAdapter The serializer to serialize an object into a string.
      * @param endpoint The workspace development endpoint, for example https://myworkspace.dev.azuresynapse.net.
      * @param apiVersion Api Version.
      */
-    public MonitoringClientImpl(
-            HttpPipeline httpPipeline, SerializerAdapter serializerAdapter, String endpoint, String apiVersion) {
+    public MonitoringClientImpl(HttpPipeline httpPipeline, SerializerAdapter serializerAdapter, String endpoint,
+        String apiVersion) {
         this.httpPipeline = httpPipeline;
         this.serializerAdapter = serializerAdapter;
         this.endpoint = endpoint;