+++ sdk/synapse/azure-analytics-synapse-spark/src/main/java/com/azure/analytics/synapse/spark/models/SparkStatementOptions.java
@@ -5,26 +5,36 @@
 package com.azure.analytics.synapse.spark.models;
 
 import com.azure.core.annotation.Fluent;
-import com.fasterxml.jackson.annotation.JsonProperty;
+import com.azure.json.JsonReader;
+import com.azure.json.JsonSerializable;
+import com.azure.json.JsonToken;
+import com.azure.json.JsonWriter;
+import java.io.IOException;
 
-/** The SparkStatementOptions model. */
+/**
+ * The SparkStatementOptions model.
+ */
 @Fluent
-public final class SparkStatementOptions {
+public final class SparkStatementOptions implements JsonSerializable<SparkStatementOptions> {
     /*
      * The code property.
      */
-    @JsonProperty(value = "code")
     private String code;
 
     /*
      * The kind property.
      */
-    @JsonProperty(value = "kind")
     private SparkStatementLanguageType kind;
 
     /**
+     * Creates an instance of SparkStatementOptions class.
+     */
+    public SparkStatementOptions() {
+    }
+
+    /**
      * Get the code property: The code property.
-     *
+     * 
      * @return the code value.
      */
     public String getCode() {
@@ -33,7 +43,7 @@ public final class SparkStatementOptions {
 
     /**
      * Set the code property: The code property.
-     *
+     * 
      * @param code the code value to set.
      * @return the SparkStatementOptions object itself.
      */
@@ -44,7 +54,7 @@ public final class SparkStatementOptions {
 
     /**
      * Get the kind property: The kind property.
-     *
+     * 
      * @return the kind value.
      */
     public SparkStatementLanguageType getKind() {
@@ -53,12 +63,51 @@ public final class SparkStatementOptions {
 
     /**
      * Set the kind property: The kind property.
-     *
+     * 
      * @param kind the kind value to set.
      * @return the SparkStatementOptions object itself.
      */
     public SparkStatementOptions setKind(SparkStatementLanguageType kind) {
         this.kind = kind;
         return this;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public JsonWriter toJson(JsonWriter jsonWriter) throws IOException {
+        jsonWriter.writeStartObject();
+        jsonWriter.writeStringField("code", this.code);
+        jsonWriter.writeStringField("kind", this.kind == null ? null : this.kind.toString());
+        return jsonWriter.writeEndObject();
+    }
+
+    /**
+     * Reads an instance of SparkStatementOptions from the JsonReader.
+     * 
+     * @param jsonReader The JsonReader being read.
+     * @return An instance of SparkStatementOptions if the JsonReader was pointing to an instance of it, or null if it
+     * was pointing to JSON null.
+     * @throws IOException If an error occurs while reading the SparkStatementOptions.
+     */
+    public static SparkStatementOptions fromJson(JsonReader jsonReader) throws IOException {
+        return jsonReader.readObject(reader -> {
+            SparkStatementOptions deserializedSparkStatementOptions = new SparkStatementOptions();
+            while (reader.nextToken() != JsonToken.END_OBJECT) {
+                String fieldName = reader.getFieldName();
+                reader.nextToken();
+
+                if ("code".equals(fieldName)) {
+                    deserializedSparkStatementOptions.code = reader.getString();
+                } else if ("kind".equals(fieldName)) {
+                    deserializedSparkStatementOptions.kind = SparkStatementLanguageType.fromString(reader.getString());
+                } else {
+                    reader.skipChildren();
+                }
+            }
+
+            return deserializedSparkStatementOptions;
+        });
     }
 }