+++ sdk/resourcemanager/azure-resourcemanager-network/CHANGELOG.md
@@ -2,13 +2,15 @@
 
 ## 2.41.0-beta.1 (Unreleased)
 
-### Features Added
-
 ### Breaking Changes
 
-### Bugs Fixed
+- Removed superclass `SubResource` from model class `FirewallPacketCaptureParameters`.
 
 ### Other Changes
+
+#### Dependency Updates
+
+- Updated `api-version` to `2024-01-01`.
 
 ## 2.40.0 (2024-06-24)